move assert to tf package
