+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/nn/tf/Assert.scala
@@ -13,10 +13,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.intel.analytics.bigdl.nn.ops
+package com.intel.analytics.bigdl.nn.tf
 
 import com.google.protobuf.ByteString
 import com.intel.analytics.bigdl.nn.abstractnn.Activity
+import com.intel.analytics.bigdl.nn.ops.Operation
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.Table
@@ -27,7 +28,8 @@ import scala.reflect.ClassTag
  * Assert will assert the first input to be true, if not, throw the message in the second
  * input. Assert has no output.
  */
-class Assert[T: ClassTag]()(implicit ev: TensorNumeric[T]) extends Operation[Table, Activity, T] {
+private[bigdl] class Assert[T: ClassTag]()
+  (implicit ev: TensorNumeric[T]) extends Operation[Table, Activity, T] {
   override def updateOutput(input: Table): Tensor[T] = {
     val predicateTensor = input(1).asInstanceOf[Tensor[Boolean]]
     val messageTensor = input(2).asInstanceOf[Tensor[ByteString]]
