+++ spark/dl/src/main/scala/com/intel/analytics/bigdl/utils/tf/loaders/Assert.scala
@@ -18,7 +18,7 @@ package com.intel.analytics.bigdl.utils.tf.loaders
 import java.nio.ByteOrder
 
 import com.intel.analytics.bigdl.Module
-import com.intel.analytics.bigdl.nn.ops.{Assert => AssertOperation}
+import com.intel.analytics.bigdl.nn.tf.{Assert => AssertOperation}
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.tensor.TensorNumericMath.TensorNumeric
 import com.intel.analytics.bigdl.utils.tf.Context
